querys arregladas

+++ src/main/java/Proyecto_Equipo_7/servicios/CalificacionServicio.java
@@ -12,7 +12,7 @@ public class CalificacionServicio {
     @Autowired
     private CalificacionRepositorio calificacionRepositorio;
     
-    public Integer promedioCalificacionesTotales(){
+    public Double promedioCalificacionesTotales(){
         
         
         return calificacionRepositorio.promedioCalificacionesTotales();
